lttng: Fix a test on Windows with file not being able to be deleted
The synthetic trace was being regenerated while it was already opened.
So it was being used in a "nested way" in the test. The test was just
trying to reopen the trace using the path so pass the path instead which
does not try to delete/regenerate the trace.

Reviewed-on: https://git.eclipse.org/r/93033
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>

+++ ctf/org.eclipse.tracecompass.ctf.core.tests/shared/org/eclipse/tracecompass/ctf/core/tests/shared/LttngTraceGenerator.java
@@ -392,6 +392,10 @@ public class LttngTraceGenerator {
 
                 @Override
                 public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
+                    // If a file failed to delete, it's more useful to throw this instead
+                    if (exc != null) {
+                        throw exc;
+                    }
                     Files.delete(dir);
                     return FileVisitResult.CONTINUE;
                 }
